+++ src/step/CopyStep.ts
@@ -4,10 +4,10 @@ import * as process from "process"
 import { IOptions } from "glob"
 import { SsgConfig } from "../SsgConfig"
 import { FileUtil } from "../util"
-import path from "path"
 
 export interface CopyStepConfig extends SsgConfig {
-  readonly copies: string[]
+  readonly sourcePatterns: string[]
+  readonly destDir: string
   readonly options?: IOptions
 }
 
@@ -20,23 +20,22 @@ export type CopyStepResult = {
  */
 export class CopyStep<C extends SsgContext = SsgContext> implements SsgStep<C, CopyStepResult> {
 
-  readonly name = 'copy';
+  readonly name = "copy"
 
   constructor(protected config: CopyStepConfig) {
   }
 
   async execute(context: SsgContext): Promise<CopyStepResult> {
-    const copies: string[] = this.config.copies
-    const destPath = this.config.getOutputPath(context)
-    const dest = path.dirname(destPath)
+    const copies: string[] = this.config.sourcePatterns
+    const dest = this.config.destDir
     try {
-      context.log('Copying to', dest, copies);
+      context.log("Copying to", dest, copies)
       const copiedFiles = await FileUtil.ssgCopy(dest, copies, this.config.options)
-      const cwd = process.cwd();
-      const files = copiedFiles.map(file => file.startsWith(cwd) ? file.substring(cwd.length + 1) : file);
-      return {files};
+      const cwd = process.cwd()
+      const files = copiedFiles.map(file => file.startsWith(cwd) ? file.substring(cwd.length + 1) : file)
+      return {files}
     } catch (e) {
-      throw Error(`Could not copy ${copies} because of ${e}`);
+      throw Error(`Could not copy ${copies} because of ${e}`)
     }
   }
 }